Include Analytics Page (#8)
### analytic.html
A new html page: this one will be used to display various data about all the ceonversations logged into the database, for only it only has 3 display options but expect more to come!
1. Summarized list of conversations (**beta**)
2. Amount of conversations per hour (**graph**)
3. Amount of conversations per day of month (**graph**)

Please take notice that the graph uses ALL the data in the base, meaning it will look for all days to gather the amount of hour and all months for the amount per day.
With this I also "modulized" the graph drawing functions, meaning they can be called in any given html page even though it's only used here for the time being.