Simplify DB Logging and Include New Graph Features (#9)
### monika.js
I tried to simplify the promise chain for the db logging function, now it at least has named functions to "easily" keep track of what does what. On top of it now it will also check for user contact information, status change (such as conversation not being successful) and user name.

### analytic.js
I reworked the interface of the analytic page to accommodate the new features, now it is possible for the user to select a variety of possible date ranges for the graph to display, and the conversations summary now only shows the last few conversations based on the date range selected.

### index.js
I reworked the index file for Monika, in short the good news is that you don't have to import the whole module anymore (in theory), the bad news is that you must call the init function now.