Include New Support Page (#10)
### pending.html
This new page allows for a human agent to load some specific information of given conversations, allowing them to get in touch with people who had a misfortunate experience with the bot. On top of that, they can register that the contact was successful or not.
The user agent may also ask the bot simple questions, and those won't be recorded to the database.

##### minor changes
1. new function to turn JS' weird DATE objects into SQL date strings.
2. actual, logical protocol
3. corrected one of the date validator regexes
4. included search options for the analytic page

That is all! I think?