Refine New Support Page (#12)
### IE SUPPORT DROPPED
With the inclusion of arrow functions, the application as a whole will ***not*** work on IE anymore. With IE's market share declining this is argumentable a good thing to do.

### pending.html / analytic.html
New feature: a human agent now has access to the "new" button, which allows them to register support that started directly with them (like say someone called to ask for information).

##### minor changes
1. "All" callback functions now use arrow functions instead for clearness
2. Restructured some if else into map lookups
3. Restructured "uneeded" functions into anonymous arrow functions
4. Included some default values for http-call methods
5. Dropped html string constructions into jquery dom builder methods, mostly

##### util.js
A new script to hold all the "common" functions across my app, it isn't complete just yet as I haven't gone through all the files just yet