Include a new interface (#14)
### 2 TRAINING DATAS, 2 INTERFACES
At first, the bot was constructed around faceb's theme and information, now with the inclusion of regius' information there are two indepedent bots, and that asks for two different interfaces too! They are pretty much the same just differently colored.

### HOME
Originally, the app's root (or home) went straight to the bot's window, now the home will open a friendly page that lets you select which of the two bots you want to chat with! There was a lot of changes to the source to accommodate this and avoid duplicated source.

##### minor changes
1. Renamed some functions (ex: loadBody to list)
2. Minor function structure improvements
3. Removed unused functions (only some of them)
4. Reorganized the chat delay into a promise chain
5. Refined the html appends/prepends calls, they look less messy now?
6. Added the option to download the current in view graph as a png file
7. Simplified a function or two or ten...
8. Refined the graph.js functions to better accommodate bigger data ranges