Migrate From MySQL to SQL Server (#16)
## H U G E C H A N G E S
### Migrated From MySQL to SQL Server
Well... I have beenr equested by the DBA guy to migrate the application to SQL Server, talk about a huge headache! But I think it was for a good reason.
The biggest change was the actual data model... At first it had domain tables so the joins used the ids like 1 for user mensagens and 2 for bot mensagens, now it uses things such as ``USU`` and ``VIR`` for that.
### Login
With the DB change comes a new feature: login! Now some of the sensitive pages such as the analytics page are locked behind a crappily done login system. Nothing is perfectly secure but I'm sure this one is far from being good, well one step at a time! The login system should identify the page you tried to access and redirectly properly after the login was successful.
### Server Log
I have created an HTML page that shows the server activity, just like the regular console, it doesn't have many features just yet but it's cool.
##### minor changes
1. Reorganized most, if not all string concats into template literals.
2. Moved the query lib from ``config.js`` to ``query-lib.js``, but that I don't have authorization to upload that one just yet.
3. General functions reorganization to fit the new model.
4. Included a new date function or two.
5. Removed some intent related functions as they became obsolete with the new model.
6. Replaced package ``promise-mysql`` with ``mssql``.
7. Minors adjustments to CSS and files.
8. Added some stuff to ``util.js``.
I thin that is all?