Update the UI and Polish API functions (#18)
### New Interface
My office randomly decided to get a rando company to make a new interface for the app. Well here it is.
### Async/Await
I have re-wrote the API to use async/await. Together with try/catch they work as syntax sugar for the gold old promises, not having to deal with a zillion nested .then()'s is GREAT.
##### minor changes
1. Polished up ``dbErr`` so it can end requests with a 500 error
2. Polished the user update function a bit with RegEx extractors
3. Changed the ``/root`` of the project to go straight to the login page
4. Kicked out more string concats in favor of template literals
5. Added a new argument to the ``dbErr`` so it can better tell at which point the catch statement triggered
6. Using the bot in itself doesn't require user login, eventually the actual user login will be integrated into the bot but for now...
7. Changed the env var ``PORT`` to ``CURRENT_PORT`` and included ``DEFAULT_PORT`` as well
There is probably more out there, but as they say: "if you don't remember then it isn't important". I guess.